Use template.WriteBuffer and move over tests
